+++ portal/src/components/Layout/components/index.ts
@@ -1,4 +1,0 @@
-export { ThemeBG } from "./ThemeBG";
-export { GitHubLinks } from "./GitHubLinks";
-export { BlogPageHeader } from "./BlogPageHeader";
-export { ComponentPageHeader } from "./ComponentPageHeader";